Made a user repository and a DDL

+++ app/src/main/java/edu/cnm/deepdive/cardcombat/model/dao/AttackCardDao.java
@@ -43,10 +43,10 @@ public interface AttackCardDao {
   Single<Integer> delete(Collection<AttackCard> attackCards);
 
   @Query("SELECT * FROM AttackCard WHERE attack_card_id = :id")
-  LiveData<AttackCard> findByAttackCardId(int id);
+  LiveData<AttackCard> findByAttackCardId(long id);
 
   @Query("SELECT * FROM AttackCard WHERE unit_card_id = :id")
-  LiveData<List<AttackCard>> findByUnitCardId(int id);
+  LiveData<List<AttackCard>> findByUnitCardId(long id);
 
 
 }
